rename `deployment.py` to `state_record.py`
